Adding vws emulator and some other stuff
To grunt, loading a grunt properties file for the grunt build (instead
of having them inline). The properties file is JSON for ease of
changing.  Considering moving the properties file and the grunt build
into its own directory... we'll see.

Adding new VWS Emulator custom grunt plugin.  Not much to say about
this, but for the devenv target, it adds a VWS emu for proxying
requests.  No more need to set up apache as a proxy (or something
similar).

Removing LESS and UGLIFY and all that from the grunt targets because I
still have to decide how to set up the build.  We have some issues to
figure out WRT how to initialize the app with config parameters before
we get into proper builds.  Soon...soon...

Changes to the application wtih regard to startup parameters are
probably broken in this build soooo yeah, wait for that.